Added admins table and model, simple static method to check if user is admin of the bot based on user id. applied to simple commands crud

+++ app/Discord/SimpleCommandCRUD.php
@@ -3,6 +3,7 @@
 namespace App\Discord;
 
 use App\Models\Command;
+use App\Models\Admin;
 use Discord\Discord;
 use Discord\Parts\Channel\Message;
 use Discord\WebSockets\Event;
@@ -16,34 +17,43 @@ class SimpleCommandCRUD
                 return;
             }
             if (str_starts_with($message->content, $bot->getPrefix() . 'addcmd ')) {
+                if (!Admin::isAdmin($message->author->id)) {
+                    return;
+                }
                 $parameters = explode(' ', $message->content);
                 if (!isset($parameters[1]) || !isset($parameters[2])) {
-                    $message->reply("Provide arguments noob");
+                    $message->channel->sendMessage("Provide arguments noob");
                 } else {
-                    $message->reply("Command saved");
+                    $message->channel->sendMessage("Command saved");
                     $command = Command::create(['trigger' => $parameters[1], 'response' => $parameters[2]]);
                     $command->save();
                     new SimpleCommand($discord, $parameters[1], $parameters[2]);
                 }
             }
 
-            if (str_starts_with($message->content,  $bot->getPrefix() . 'delcmd ')) {
+            if (str_starts_with($message->content, $bot->getPrefix() . 'delcmd ')) {
+                if (!Admin::isAdmin($message->author->id)) {
+                    return;
+                }
                 $parameters = explode(' ', $message->content);
                 if (!isset($parameters[1])) {
-                    $message->reply("Provide arguments noob");
+                    $message->channel->sendMessage("Provide arguments noob");
                 } else {
                     Command::where(['trigger' => $parameters[1]])->delete();
-                    $message->reply("Command deleted");
+                    $message->channel->sendMessage("Command deleted");
                 }
 
             }
 
-            if ($message->content ==  $bot->getPrefix() . 'commands') {
+            if ($message->content == $bot->getPrefix() . 'commands') {
+                if (!Admin::isAdmin($message->author->id)) {
+                    return;
+                }
                 $commands = '';
                 foreach (Command::all() as $command) {
                     $commands .= $command->trigger . ' ';
                 }
-                $message->reply($commands);
+                $message->channel->sendMessage($commands);
             }
 
         });
